+++ UnityProject/Assets/Scripts/Messages/Server/UpdateUIMessage.cs
@@ -8,7 +8,7 @@ using UnityEngine.Networking;
 public class UpdateUIMessage : ServerMessage
 {
 	public static short MessageType = (short)MessageTypes.UpdateUIMessage;
-	public int CurHealth;
+	public float CurHealth;
 	public bool ShowOxyWarning;
 
 	public override IEnumerator Process()
@@ -27,7 +27,7 @@ public class UpdateUIMessage : ServerMessage
 	/// Update health meter for given player
 	/// <param name="recipient">Recipient.</param>
 	/// <param name="cHealth">Current server health.</param>
-	public static UpdateUIMessage SendHealth(GameObject recipient, int cHealth)
+	public static UpdateUIMessage SendHealth(GameObject recipient, float cHealth)
 	{
 		UpdateUIMessage msg = new UpdateUIMessage
 		{
